Refactor: NodeWeight extracted to partial class.
